Stats + lv up
working stat system and a leveling syste.

fixed issues with damaging enemies (not all issues)

+++ Games Fleadh Maze Game/Assets/Art/Enemies/Skeleton W_Sword/swordDamage.cs
@@ -4,7 +4,7 @@ using UnityEngine;
 
 public class swordDamage : MonoBehaviour {
 
-	public float dmg;
+	public int dmg;
 
 	void OnTriggerEnter(Collider other){
 		StartCoroutine ("Collider");
